server/requirements.txt
fastapi
uvicorn
pydantic
openai
openenv-core
